Moved constants to the domain module
