+++ src/components/analyzers/basic-analyzer/front-view/front-view.jsx
@@ -1,5 +1,5 @@
 import drawingStyles from "../../drawing.module.css";
-import { Buttons, Joysticks } from "../../../../constants";
+import { Buttons, Joysticks } from "../../../../domain";
 import { useButton, useJoystick } from "../../../../hooks";
 import { Joystick } from "./joystick";
 import { DirectionalPad } from "./directional-pad";
