--- src/constants/index.js
@@ -1,3 +1,0 @@
-export { Buttons } from "./buttons";
-export { Joysticks } from "./joysticks";
-export { Analyzers } from "./analyzers";
