+++ src/domain/index.ts
@@ -1,1 +1,4 @@
+export { Analyzers } from "./analyzers";
+export { Buttons } from "./buttons";
 export { ControllerType } from "./controller-type";
+export { Joysticks } from "./joysticks";
